- Fix some compiler warnings.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@17817 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ imagery/i.class/outline.c
@@ -156,7 +156,7 @@ int outline (void)
   /* sort the edge points by row and then by col */
 
   Menu_msg("Sorting...");
-  qsort (P, PN, sizeof(POINT), edge_order);
+  qsort (P, (size_t)PN, sizeof(POINT), edge_order);
 
   Menu_msg("");
   return (1);
